// src/app/models/social-ethical.model.ts
export interface SocialEthicalData {
    laborRightsPolicy: boolean;
    laborRightsPolicyPublic: boolean;
    humanRightsDueDiligence: boolean;
    humanRightsDueDiligenceFrequency: string;
    collectiveBargainingPercentage: number;
    employeeGrievances: number;
    resolvedEmployeeGrievances: number;
    grievanceMechanism: boolean;
    communityInvestment: number;
    communityInvestmentPercentage: number;
    communityInitiatives: number;
    volunteerHours: number;
    socialImpactDescription: string;
  }